requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
drf-yasg
PyYAML
